fix: resolve the purchase order flow
