+++ packages/web-service/src/pages/contact/invoice/__tests__/common.spec.js
@@ -1,5 +1,7 @@
-describe('the getData function', () => {
-  it('calls the APIs', async () => {
+describe('the common functions', () => {
+  beforeEach(() => jest.resetModules())
+
+  it('getData calls the APIs', async () => {
     const mockFn = jest.fn()
     jest.doMock('../../../../services/api-requests.js', () => ({
       APIRequests: {
@@ -33,5 +35,45 @@ describe('the getData function', () => {
     const { generateOutput } = await import('../common.js')
     await generateOutput(request)
     expect(mockFn).toHaveBeenCalledTimes(3)
+  })
+
+  it('purchaseOrderCompletion returns the user to CYA if the user has submitted a referenceOrPurchaseOrderNumber', async () => {
+    jest.doMock('../../../../services/api-requests.js', () => ({
+      APIRequests: {
+        APPLICATION: {
+          getById: () => {
+            return {
+              referenceOrPurchaseOrderNumber: '123abc'
+            }
+          }
+        }
+      }
+    }))
+    const request = {
+      cache: () => ({
+        getData: () => ({ applicationId: '68855554-59ed-ec11-bb3c-000d3a0cee24' })
+      })
+    }
+    const { purchaseOrderCompletion } = await import('../common.js')
+    expect(await purchaseOrderCompletion(request)).toEqual('/invoice-check-answers')
+  })
+
+  it('purchaseOrderCompletion returns the user to invoice-purchase-order if the user has not submitted a referenceOrPurchaseOrderNumber', async () => {
+    jest.doMock('../../../../services/api-requests.js', () => ({
+      APIRequests: {
+        APPLICATION: {
+          getById: () => {
+            return {}
+          }
+        }
+      }
+    }))
+    const request = {
+      cache: () => ({
+        getData: () => ({ applicationId: '68855554-59ed-ec11-bb3c-000d3a0cee24' })
+      })
+    }
+    const { purchaseOrderCompletion } = await import('../common.js')
+    expect(await purchaseOrderCompletion(request)).toEqual('/invoice-purchase-order')
   })
 })
