+++ packages/web-service/src/pages/contact/invoice/invoice-address-form.js
@@ -5,19 +5,9 @@ import { getAddressFormData, setAddressFormData } from '../common/address-form/a
 import { AccountRoles, ContactRoles } from '../common/contact-roles.js'
 import { checkHasContact } from '../common/common-handler.js'
 import { checkApplication } from '../../common/check-application.js'
-import { APIRequests } from '../../../services/api-requests.js'
+import { purchaseOrderCompletion } from './common.js'
 
-const { ADDRESS_FORM, CHECK_ANSWERS, RESPONSIBLE, PURCHASE_ORDER } = contactURIs.INVOICE_PAYER
-
-export const completion = async request => {
-  const { applicationId } = await request.cache().getData()
-  const applicationData = await APIRequests.APPLICATION.getById(applicationId)
-  if (!applicationData.referenceOrPurchaseOrderNumber) {
-    return PURCHASE_ORDER.uri
-  } else {
-    return CHECK_ANSWERS.uri
-  }
-}
+const { ADDRESS_FORM, RESPONSIBLE } = contactURIs.INVOICE_PAYER
 
 export const invoiceAddressForm = addressFormPage({
   page: ADDRESS_FORM.page,
@@ -25,5 +15,5 @@ export const invoiceAddressForm = addressFormPage({
   checkData: [checkApplication, checkHasContact(ContactRoles.PAYER, RESPONSIBLE)],
   getData: getAddressFormData(ContactRoles.PAYER, AccountRoles.PAYER_ORGANISATION),
   setData: setAddressFormData(ContactRoles.PAYER, AccountRoles.PAYER_ORGANISATION, contactURIs.INVOICE_PAYER),
-  completion
+  completion: purchaseOrderCompletion
 })
